requirements.txt
streamlit
requests
pandas
requests
beautifulsoup4
imdbpy
youtube-search-python
streamlit_player